make Grid structure variable
I would like to dynamically set Grid variables like the FOCAL_LENGTH and
the *_RING_RADIUS. This commit introduces the `Grid::Structure`
struct, which holds those variables.

For the pipeline use-case this commit adds an overhead of
an incrementation of a shared pointer, the default structure and
`_coordinates3D` are cached.